RedditAPI/requirements.txt

praw==7.8.1
requests>=2.32.0
python-dotenv>=1.0.1
pydantic>=2.8.2
PyYAML>=6.0.2
typer>=0.12.5
rich>=13.7.1
pandas>=2.2.2
matplotlib>=3.9.0
